Revert "enums: fixed conversion warning (#4025)"
This reverts commit 60b469cd6af14cb4692bc30a58f9709ed16ea611.

+++ src/tile.h
@@ -138,7 +138,7 @@ class TileItemVector : private ItemVector
 			}
 			return *getBeginDownItem();
 		}
-		void addDownItemCount(uint16_t increment) {
+		void addDownItemCount(int32_t increment) {
 			downItemCount += increment;
 		}
